+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 **(Next Version) Changes:**
 
-* Fixed effects not activating in stages that pause the run timer.
+* Effects will now activate in stages with the run timer paused
 * Fixed effects being able to activate while the game is paused if the time between effects config value is changed
 
 * Added Effect: Teleport to Random Location